telepoop-backend/requirements.txt
boto3
chalice
pydantic
pynamodb
pyjwt